Fix GRPO activation checkpointing weight sync
