+++ open_instruct/grpo_callbacks.py
@@ -7,7 +7,6 @@ These callbacks handle:
 """
 
 import contextlib
-import re
 import time
 from collections.abc import Callable
 from dataclasses import dataclass, field
@@ -26,42 +25,9 @@ from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 
 from open_instruct import data_loader as data_loader_lib
 from open_instruct import grpo_utils, logger_utils, utils, vllm_utils
+from open_instruct.grpo_name_mapping import olmo_core_to_hf_name
 
 logger = logger_utils.setup_logger(__name__)
-
-_BLOCK_PATTERN = re.compile(r"blocks\.(\d+)\.(.*)")
-_OLMO_CORE_TO_HF_LAYER_MAPPINGS = {
-    "attention.w_q.weight": "self_attn.q_proj.weight",
-    "attention.w_k.weight": "self_attn.k_proj.weight",
-    "attention.w_v.weight": "self_attn.v_proj.weight",
-    "attention.w_out.weight": "self_attn.o_proj.weight",
-    "attention.q_norm.weight": "self_attn.q_norm.weight",
-    "attention.k_norm.weight": "self_attn.k_norm.weight",
-    "feed_forward.w1.weight": "mlp.gate_proj.weight",
-    "feed_forward.w2.weight": "mlp.down_proj.weight",
-    "feed_forward.w3.weight": "mlp.up_proj.weight",
-    "attention_norm.weight": "input_layernorm.weight",
-    "feed_forward_norm.weight": "post_attention_layernorm.weight",
-}
-
-
-def olmo_core_to_hf_name(name: str) -> str:
-    """Convert OLMo-core parameter name to HuggingFace format for Qwen3/LLaMA models."""
-    if name == "embeddings.weight":
-        return "model.embed_tokens.weight"
-    if name == "lm_head.norm.weight":
-        return "model.norm.weight"
-    if name == "lm_head.w_out.weight":
-        return "lm_head.weight"
-
-    layer_match = _BLOCK_PATTERN.match(name)
-    if layer_match:
-        layer_idx = layer_match.group(1)
-        rest = layer_match.group(2)
-        if rest in _OLMO_CORE_TO_HF_LAYER_MAPPINGS:
-            return f"model.layers.{layer_idx}.{_OLMO_CORE_TO_HF_LAYER_MAPPINGS[rest]}"
-
-    return name
 
 
 @dataclass
